tools/WebFileSync/requirements.txt
FreeSimpleGUI>=5.0
python-dateutil>=2.8
